requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.1.2
django-restframework==0.0.1
pycodestyle==2.9.1
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.5
